+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/teleop/FieldCentricTeleOp.java
@@ -4,7 +4,7 @@ import com.qualcomm.hardware.sparkfun.SparkFunOTOS;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.teamcode.subassemblies.Follower;
+import org.firstinspires.ftc.teamcode.subassemblies.autonomous.Follower;
 
 @TeleOp(group = "exploratory")
 public class FieldCentricTeleOp extends LinearOpMode {
