+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subassemblies/autonomous/localizers/LimelightCam.java
@@ -1,10 +1,9 @@
-package org.firstinspires.ftc.teamcode.subassemblies.odometry;
+package org.firstinspires.ftc.teamcode.subassemblies.autonomous.localizers;
 
 import com.qualcomm.hardware.limelightvision.LLResult;
 import com.qualcomm.hardware.limelightvision.Limelight3A;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.util.Localizer;
 import org.firstinspires.ftc.teamcode.util.Pose;
 
 public class LimelightCam extends Localizer {
