+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subassemblies/autonomous/localizers/PinpointOdo.java
@@ -1,8 +1,6 @@
-package org.firstinspires.ftc.teamcode.subassemblies.odometry;
+package org.firstinspires.ftc.teamcode.subassemblies.autonomous.localizers;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-import org.firstinspires.ftc.teamcode.util.Localizer;
 
 public class PinpointOdo extends Localizer {
