+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subassemblies/autonomous/localizers/SparkFunOpticalOdo.java
@@ -1,11 +1,10 @@
-package org.firstinspires.ftc.teamcode.subassemblies.odometry;
+package org.firstinspires.ftc.teamcode.subassemblies.autonomous.localizers;
 
 import com.qualcomm.hardware.sparkfun.SparkFunOTOS;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
-import org.firstinspires.ftc.teamcode.util.Localizer;
 import org.firstinspires.ftc.teamcode.util.Pose;
 
 public class SparkFunOpticalOdo extends Localizer {
